spike: breaking circular dependency
